chore: rename files for consistency
We have some kebab-case and some camelCase files, switch all to
kebab-case for consistency with other Helia projects.
