+++ package.json
@@ -4,6 +4,7 @@
   "description": "Helia in Docker Container",
   "main": "src/index.ts",
   "scripts": {
+    "clean": "aegir clean dist playwright-report test",
     "lint": "aegir lint",
     "build": "aegir build --bundle false",
     "dep-check": "aegir dep-check",
@@ -76,14 +77,14 @@
   "dependencies": {
     "@chainsafe/libp2p-noise": "^15.0.0",
     "@chainsafe/libp2p-yamux": "^6.0.2",
-    "@fastify/compress": "^6.5.0",
-    "@fastify/cors": "^8.5.0",
+    "@fastify/compress": "^7.0.0",
+    "@fastify/cors": "^9.0.1",
     "@helia/block-brokers": "^2.0.2",
     "@helia/delegated-routing-v1-http-api-client": "^3.0.0",
-    "@helia/http": "^1.0.2",
+    "@helia/http": "^1.0.3",
     "@helia/interface": "^4.0.1",
     "@helia/routers": "^1.0.1",
-    "@helia/utils": "^0.0.2",
+    "@helia/utils": "^0.1.0",
     "@helia/verified-fetch": "^1.1.2",
     "@libp2p/bootstrap": "^10.0.16",
     "@libp2p/circuit-relay-v2": "^1.0.16",
@@ -102,9 +103,9 @@
     "@sgtpooki/file-type": "^1.0.1",
     "blockstore-level": "^1.1.8",
     "datastore-level": "^10.1.7",
-    "fastify": "4.25.2",
-    "fastify-metrics": "10.3.3",
-    "helia": "^4.0.2",
+    "fastify": "^4.26.2",
+    "fastify-metrics": "^11.0.0",
+    "helia": "^4.1.0",
     "ipns": "^9.0.0",
     "libp2p": "^1.2.4",
     "multiformats": "^13.0.1",
